+++ include/render/materials/ReflectiveMaterial.h
@@ -133,7 +133,7 @@ namespace render {
                                     render::State& state) const override;
 
     bool requiresWhittedPacketHitRefinement() const override {
-      return true;
+      return false;
     }
 
     const char* whittedPacketHitRefinementLabel() const override {