+++ include/render/materials/TransparentMaterial.h
@@ -174,7 +174,7 @@ namespace render {
                                     render::State& state) const override;
 
     bool requiresWhittedPacketHitRefinement() const override {
-      return true;
+      return false;
     }
 
     const char* whittedPacketHitRefinementLabel() const override {